+++ frontend/src/components/session/SessionManager.tsx
@@ -13,6 +13,7 @@ import {
 } from 'lucide-react';
 import { Button } from '@/components/ui/Button';
 import { Card, CardContent } from '@/components/ui/Card';
+import { formatDuration } from '@/lib/utils/time';
 
 interface SessionManagerProps {
   isRecording: boolean;
@@ -39,16 +40,7 @@ export function SessionManager({
   onSaveSession,
   onExportSession
 }: SessionManagerProps) {
-  const formatDuration = (seconds: number) => {
-    const hours = Math.floor(seconds / 3600);
-    const minutes = Math.floor((seconds % 3600) / 60);
-    const secs = seconds % 60;
-    
-    if (hours > 0) {
-      return `${hours}:${minutes.toString().padStart(2, '0')}:${secs.toString().padStart(2, '0')}`;
-    }
-    return `${minutes}:${secs.toString().padStart(2, '0')}`;
-  };
+
 
   const calculateTalkRatio = () => {
     // This would be calculated based on actual speaker distribution
