+++ frontend/src/lib/conversation/stateUtils.ts
@@ -54,14 +54,6 @@ export const getStateInfo = (state: ConversationState): ConversationStateInfo =>
   }
 };
 
-/**
- * Format duration from seconds to MM:SS format.
- */
-export const formatDuration = (seconds: number): string => {
-  const mins = Math.floor(seconds / 60);
-  const secs = seconds % 60;
-  return `${mins.toString().padStart(2, '0')}:${secs.toString().padStart(2, '0')}`;
-};
 
 /**
  * Update talk statistics based on new transcript text.
